Update network-protection.md
The kusto query in this article is not showing Network protection events neither audited nor blocked because it recommend to query from "DeviceNetworkEvents" but network event are stored in "DeviceEvents" scheme. Therefore KQL query as such will show only "connection success" and those are quite irrelevant in case you hunt for Network Protection events that has been audited or blocked.